# openapi/driveit.yaml
openapi: "3.0.0"
info:
  description: "Bike autostation service"
  version: "1.0.0"
  title: "Drive it"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "localhost:8080/api/v1"
...
paths:
...
  /bikes:
    post:
      tags: 
      - "bikes"
      summary: "Register a new bike"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BikeTO'
      responses:
        '202':
          description: "Created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseTO"
        '400':
          description: "Bad request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseTO"
  /bikes/{bikeId}:
    get:
      tags: 
      - "bikes"
      summary: "Returns the specified user"
      parameters:
      - name: bikeId
        in: path
        required: true
        description: "The user identifier"
        schema:
          type: "string"
          pattern: "[a-zA-Z0-9- _]$"
          maxLength: 200
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserTO"
        '400':
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseTO"
        '404':
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseTO"
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseTO"
...
components:
  schemas: 
    UserTO:
      type: "object"
      required: 
      - "name"
      - "surname"
      properties:
        id:
          type: integer
          format: "int64"
          minimum: 1
          maximum: 1000000
        name:
          type: string
          pattern: "[a-zA-Z0-9- _]$"
          maxLength: 100
        surname:
          type: string
          pattern: "[a-zA-Z0-9- _]$"
          maxLength: 100
        usingBike:
          type: boolean
        totalHours:
          type: number
          format: decimal
          minimum: 0
          maximum: 100000000000.00000000
        status:
          type: boolean
    BikeTO:
      type: "object"
      required: 
      - "serie"
      - "station"
      properties:
        serie:
          type: string
          pattern: "[a-zA-Z0-9- _]$"
          maxLength: 200
        station:
          type: integer
          format: "int64"
          minimum: 1
          maximum: 1000000
...
    ResponseTO:
      type: "object"
      required:
      - "statusCode"
      - "message"
      properties:
        statusCode:
          type: string
          pattern: "[a-zA-Z0-9- _]$"
          maxLength: 200
        message:
          type: string
          pattern: "[a-zA-Z0-9- _]$"
          maxLength: 200
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []